pyproject.toml
[project]
name = "audio-process"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Baurzhan Beglerov", email = "bbeglerov@icloud.com" }
]
requires-python = ">=3.14"
dependencies = [
    "cloudpickle>=3.1.2",
    "python-dotenv>=1.0.0",
    "cmake>=4.3.1",
    "datasets>=4.8.4",
    "editdistance>=0.8.1",
    "fiddle>=0.3.0",
    "google-genai>=1.71.0",
    "google-cloud-speech>=2.31.1",
    "huggingface-hub>=1.9.2",
    "hydra-core>=1.3.2",
    "ipython>=9.12.0",
    "jiwer>=4.0.0",
    "kaldialign>=0.9.3",
    "lhotse>=1.32.2",
    "librosa>=0.11.0",
    "openai>=2.30.0",
    "openai-whisper>=20250625",
    "overrides>=7.3.0",
    "pyannote-audio>=4.0.4",
    "pyannote-core>=6.0.1",
    "python-dateutil>=2.9.0.post0",
    "pytorch-lightning<=2.2.0",
    "resampy>=0.4.3",
    "ruamel-yaml>=0.19.1",
    "scikit-learn>=1.8.0",
    "sentencepiece>=0.2.1",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "text-unidecode>=1.3",
    "torch>=2.11.0",
    "torchmetrics>=1.9.0",
    "transformers>=5.5.0",
    "webdataset>=1.0.2",
    "wget>=3.2",
    "wrapt>=2.1.2",
    "pydantic>=2.12.5",
    "dagster>=1.13.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
